+++ CHANGES.md
@@ -5,7 +5,7 @@ Features ✨:
  -
 
 Improvements 🙌:
- -
+ - New wording for notice when current user is the sender
 
 Bugfix 🐛:
  - Switch theme is not fully taken into account without restarting the app